agents/coach-companion-agentcore/requirements.txt
# Strands SDK for agent orchestration
strands-agents>=0.1.0

# AgentCore SDK
bedrock-agentcore>=0.1.0

# AWS SDK
boto3>=1.34.0
